Fill db schema failed_signins field too small
Changed failed_signins.attempted from smallint to int.

+++ etc/schema.sql
@@ -106,7 +106,7 @@ CREATE TABLE `failed_signins` (
   `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
   `usersId` int(10) unsigned DEFAULT NULL,
   `ipAddress` char(15) NOT NULL,
-  `attempted` smallint(5) unsigned NOT NULL,
+  `attempted` int(10) unsigned NOT NULL,
   PRIMARY KEY (`id`),
   KEY `usersId` (`usersId`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
